+++ src/ReadoutUtils.h
@@ -45,6 +45,10 @@ int getKeyValuePairsFromString(const std::string& input, std::map<std::string, s
 // returns 0 on success, -1 on error
 int getIntegerListFromString(const std::string& input, std::vector<int>& output);
 
+// parse a string of coma-separated strings into a vector
+// returns 0 on success, -1 on error
+int getListFromString(const std::string& input, std::vector<std::string>& output);
+
 // function to convert a value in bytes to a prefixed number 3+3 digits
 // suffix is the "base unit" to add after calculated prefix, e.g. Byte-> kBytes
 std::string NumberOfBytesToString(double value, const char* suffix, int base = 1024);
@@ -56,15 +60,15 @@ int getProcessStats(double& uTime, double& sTime);
 typedef uint32_t tRunNumber;
 typedef uint32_t tTimeframeId;
 
-// function to retrieve amount of free memory on the system
+// function to retrieve some memory statistics on the system
 // Works only when /proc/meminfo available
+// Look for entry corresponding to provided keyword (eg: MemFree, MemAvailable)
 // returns 0 on success, -1 on error
-int getStatsFreeMemory(unsigned long long &freeBytes);
+int getStatsMemory(unsigned long long &freeBytes, const std::string& keyword);
 
-// function to retrieve amount of free memory on the system
-// Works only when /dev/shm available
+// function to retrieve amount of free area for given path on the filesystem
 // returns 0 on success, -1 on error
-int getStatsFreeSHM(unsigned long long &freeBytes);
+int getStatsFilesystem(unsigned long long &freeBytes, const std::string& path);
 
 // end of _READOUTUTILS_H
 #endif