+++ doc/releaseNotes.md
@@ -354,3 +354,8 @@ This file describes the main feature changes for each readout.exe released versi
 - equipment-cruemulator: TF id extracted from trigger counters (single timer source for improved coherency).
 - Memory allocation policy updated: all readout memory is locked (RAM only, can not be swapped). A warning is reported if not.
 - consumer-FMQchannel: checks are done before FMQ shared memory region is created, to avoid going in a state with over-committed memory (no checks done in FMQ library about the validity of the region created, which can cause severe crash when trying to access it). Both /proc/meminfo (MemFree) and /dev/shm (if using shmem transport type) should report enough available memory before proceeding. Memory is also immediately locked and zeroed to avoid later crashes.
+
+## Next version
+- Updated configuration parameters:
+  - added readout.disableTimefarmes:  when set, all timeframe-related features are disabled (STF slicing, TF rate limits, etc). All data are tagged with TF id = 0. To be used for some calibration runs not using a central trigger clock.
+  - added consumer-FMQchannel.checkResources: controls which resources are checked for fitting unmanaged region. This is a comma-separated list of items to be checked. By default, no checks are done. Recommended value: /dev/shm, MemAvailable.